Added better serialization for embed language
The serialization for embed languages was quite poor, the idea was to
get the first version out quickly. This meant dumping the JSON as a
string and using serde-runtime to create the structs needed. This means
that on boot a JSON would be read off the binary.

This branch improves that by creating the pub without any serde runtime.
Instead of dumping a bunch of JSON code is generated off the JSON that
is stored ad the training stage.